# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/login": {
      "post": {
        "operationId": "UserController_loginOrCreate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/verify": {
      "post": {
        "operationId": "UserController_verify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPhoneNumberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreatedResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/user": {
      "post": {
        "operationId": "UserController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      },
      "get": {
        "operationId": "UserController_getCurrentUserProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/task/create": {
      "post": {
        "operationId": "TaskController_createTask",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Task"
        ]
      }
    },
    "/api/task/update": {
      "post": {
        "operationId": "TaskController_updateTaskStatus",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCompletionStatus"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Task"
        ]
      }
    },
    "/api/task/edit": {
      "post": {
        "operationId": "TaskController_editTask",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Task"
        ]
      }
    },
    "/api/task/get": {
      "get": {
        "operationId": "TaskController_getAllTasks",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apartmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Task"
        ]
      }
    },
    "/api/task/get-completed": {
      "get": {
        "operationId": "TaskController_getCompletedTasks",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Task"
        ]
      }
    },
...
  },
  "info": {
    "title": "Komuna",
    "description": "Komuna API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "SplitType": {
        "type": "string",
        "enum": [
          "EQUAL",
          "NUMBER",
          "PERCENTAGE"
        ]
      },
      "Expense": {
        "type": "object",
        "properties": {
          "expenseId": {
            "type": "string",
            "description": "Unique identifier for the expense"
          },
          "apartmentId": {
            "type": "string",
            "description": "ID of the apartment associated with the expense"
          },
          "description": {
            "type": "string",
            "description": "Description of the expense"
          },
          "amount": {
            "type": "number",
            "description": "Amount of the expense"
          },
          "paidById": {
            "type": "string",
            "description": "ID of the user who paid for the expense"
          },
          "splits": {
            "type": "object",
            "description": "Splits of the expense among users",
            "example": {
              "user1": 50,
              "user2": 50
            },
            "additionalProperties": {
              "type": "number"
            }
          },
          "splitType": {
            "description": "Type of split for the expense",
            "allOf": [
              {
                "$ref": "#/components/schemas/SplitType"
              }
            ]
          },
          "paidByUser": {
            "description": "User who paid for the expense",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date of the expense"
          }
        },
        "required": [
          "expenseId",
          "apartmentId",
          "description",
          "amount",
          "paidById",
          "splits",
          "splitType",
          "paidByUser",
          "createdAt"
        ]
      },
      "Payment": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string",
            "description": "Unique identifier for the payment"
          },
          "apartmentId": {
            "type": "string",
            "description": "ID of the apartment where the payment is made"
          },
          "fromId": {
            "type": "string",
            "description": "ID of the user making the payment"
          },
          "toId": {
            "type": "string",
            "description": "ID of the user receiving the payment"
          },
          "amount": {
            "type": "number",
            "description": "Amount of the payment"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the payment was created"
          }
        },
        "required": [
          "paymentId",
          "apartmentId",
          "fromId",
          "toId",
          "amount",
          "createdAt"
        ]
      },
      "IncidentUrgency": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "EXTREME"
        ]
      },
      "IncidentStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "IN_PROGRESS",
          "SOLVED"
        ]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "commentId": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "incidentId": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "user": {
            "type": "string"
          }
        },
        "required": [
          "commentId",
          "message",
          "userId",
          "createdAt",
          "incidentId",
          "images",
          "user"
        ]
      },
      "Incident": {
        "type": "object",
        "properties": {
          "urgencyLevel": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IncidentUrgency"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IncidentStatus"
              }
            ]
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "reporter": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "urgencyLevel",
          "status",
          "comments",
          "reporter"
        ]
      },
      "ShoppingTemplate": {
        "type": "object",
        "properties": {}
      },
      "ContextType": {
        "type": "string",
        "enum": [
          "APARTMENT",
          "USER"
        ]
      },
      "ShoppingListItemWithIdDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "isPurchased": {
            "type": "boolean"
          },
          "category": {
            "type": "string"
          },
          "isUrgent": {
            "type": "boolean"
          },
          "amount": {
            "type": "number"
          },
          "creatorId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "itemId": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "isPurchased",
          "isUrgent",
          "amount",
          "itemId"
        ]
      },
      "ShoppingList": {
        "type": "object",
        "properties": {
          "contextType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContextType"
              }
            ]
          },
          "contextId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShoppingListItemWithIdDto"
            }
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "contextType",
          "contextId",
          "items",
          "updatedAt",
          "createdAt"
        ]
      },
      "Apartment": {
        "type": "object",
        "properties": {
          "apartmentId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "landlordCode": {
            "type": "string",
            "description": "Unique code to join the apartment as a landlord. Can be NULL when the apartment already has a landlord"
          },
          "roommateCode": {
            "type": "string",
            "description": "Unique code to join the apartment as a roommate. NULL in case the apartment doesn't allow new residents"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "landlordUserId": {
            "type": "string",
            "description": "Landlord User ID"
          },
          "landlord": {
            "description": "Landlord of the apartment (Relation)",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "contractEndDate": {
            "format": "date-time",
            "type": "string",
            "description": "Apartment contract end date"
          },
          "rent": {
            "type": "number",
            "description": "Apartment rent"
          },
          "billsDetails": {
            "type": "object",
            "description": "Bills payment details"
          },
          "houseCommitteeRent": {
            "type": "number",
            "description": "Monthly house committee rent"
          },
          "houseCommitteePayerUser": {
            "description": "User ID of the house committee payer. NULL if it's split equally",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "residents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserApartment"
            }
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserApartment"
            }
          },
          "expenses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Expense"
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Payment"
            }
          },
          "incidents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Incident"
            }
          },
          "shoppingTemplates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShoppingTemplate"
            }
          },
          "shoppingLists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShoppingList"
            }
          },
          "createdAt": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          }
        },
        "required": [
          "apartmentId",
          "name",
          "image",
          "landlordCode",
          "roommateCode",
          "address",
          "city",
          "residents",
          "tasks",
          "expenses",
          "payments",
          "incidents",
          "shoppingTemplates",
          "shoppingLists",
          "createdAt"
        ]
      },
      "UserApartment": {
        "type": "object",
        "properties": {
          "apartmentId": {
            "type": "string",
            "description": "The ID of the apartment where the user resides"
          },
          "userId": {
            "type": "string",
            "description": "The ID of the user who is a resident in the apartment"
          },
          "rent": {
            "type": "number",
            "description": "The amount of rent the user pays for the apartment"
          },
          "role": {
            "type": "string",
            "description": "The role of the user in the apartment",
            "enum": [
              "LANDLORD",
              "ROOMMATE"
            ]
          },
          "apartment": {
            "description": "The apartment where the user resides",
            "allOf": [
              {
                "$ref": "#/components/schemas/Apartment"
              }
            ]
          },
          "user": {
            "description": "The user who is a resident in the apartment",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "payableByUser": {
            "description": "The user who is responsible for paying rent",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date when the user-apartment relationship was created"
          }
        },
        "required": [
          "apartmentId",
          "userId",
          "rent",
          "role",
          "apartment",
          "user",
          "payableByUser",
          "createdAt"
        ]
      },
      "DebtEdge": {
        "type": "object",
        "properties": {
          "debtId": {
            "type": "string",
            "description": "Unique identifier for the debt"
          },
          "apartmentId": {
            "type": "string",
            "description": "ID of the apartment this debt belongs to"
          },
          "fromId": {
            "type": "string",
            "description": "ID of the user who owes the debt"
          },
          "toId": {
            "type": "string",
            "description": "ID of the user who is owed the debt"
          },
          "amount": {
            "type": "number",
            "description": "Amount of the debt"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date of the debt"
          },
          "fromUser": {
            "description": "User who owes the debt",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "toUser": {
            "description": "User who is owed the debt",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          }
        },
        "required": [
          "debtId",
          "apartmentId",
          "fromId",
          "toId",
          "amount",
          "updatedAt",
          "fromUser",
          "toUser"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "description": "URL to user's profile image"
          },
          "apartments": {
            "description": "User's apartments",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserApartment"
            }
          },
          "landlordApartments": {
            "description": "User's apartments as landlord",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Apartment"
            }
          },
          "payableRents": {
            "description": "User's payable rents",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserApartment"
            }
          },
          "expenses": {
            "description": "User's expenses",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Expense"
            }
          },
          "debts": {
            "description": "User's debts",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DebtEdge"
            }
          },
          "credits": {
            "description": "User's credits",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DebtEdge"
            }
          },
          "comments": {
            "description": "The comments made by the user",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "incidents": {
            "description": "The incidents reported by the user",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Incident"
            }
          }
        },
        "required": [
          "userId",
          "firstName",
          "phoneNumber",
          "lastName",
          "apartments",
          "landlordApartments",
          "payableRents",
          "expenses",
          "debts",
          "credits",
          "comments",
          "incidents"
        ]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "User's phone number for login"
          }
        },
        "required": [
          "phoneNumber"
        ]
      },
      "VerifyPhoneNumberDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "User's phone number to verify"
          },
          "code": {
            "type": "string",
            "description": "Verification code sent to the phone number"
          }
        },
        "required": [
          "phoneNumber",
          "code"
        ]
      },
      "UserCreatedResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "isUser": {
            "type": "boolean"
          }
        },
        "required": [
          "user",
          "isUser"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "image": {
            "type": "string"
          }
        },
        "required": [
          "phoneNumber",
          "firstName",
          "lastName",
          "image"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          }
        },
        "required": [
          "user"
        ]
      },
...
    }
  }
}